requirements.txt
Flask
pdfminer.six
fuzzywuzzy
scikit-learn
joblib
numpy
